moved login into routes folder
